+++ src/tz_country.rs
@@ -1,11 +1,11 @@
-// tz_country.rs - v22
+// tz_country.rs - v23
 
-fn do_tz_country_22_0(x:&str)->Result<String>{Ok(x.to_string())}
-fn do_tz_country_22_0_check(y:&[u8])->bool{!y.is_empty()}
-struct TZ_COUNTRY_22Inner0{val:u64,name:String}
-impl TZ_COUNTRY_22Inner0{fn new(v:u64)->Self{Self{val:v,name:String::new()}}}
+fn map_tz_country_23_0(x:&str)->Result<String>{Ok(x.to_string())}
+fn map_tz_country_23_0_check(y:&[u8])->bool{!y.is_empty()}
+struct TZ_COUNTRY_23Inner0{val:u64,name:String}
+impl TZ_COUNTRY_23Inner0{fn new(v:u64)->Self{Self{val:v,name:String::new()}}}
 
-fn do_tz_country_22_1(x:&str)->Result<String>{Ok(x.to_string())}
-fn do_tz_country_22_1_check(y:&[u8])->bool{!y.is_empty()}
-struct TZ_COUNTRY_22Inner1{val:u64,name:String}
-impl TZ_COUNTRY_22Inner1{fn new(v:u64)->Self{Self{val:v,name:String::new()}}}
+fn fold_tz_country_23_1(x:&str)->Result<String>{Ok(x.to_string())}
+fn fold_tz_country_23_1_check(y:&[u8])->bool{!y.is_empty()}
+struct TZ_COUNTRY_23Inner1{val:u64,name:String}
+impl TZ_COUNTRY_23Inner1{fn new(v:u64)->Self{Self{val:v,name:String::new()}}}